Move constants file to base package
