+++ src/main/java/org/opensrp/web/Constants.java
@@ -1,0 +1,9 @@
+package org.opensrp.web;
+
+public interface Constants {
+
+    String LIMIT = "limit";
+
+    int DEFAULT_LIMIT = 25;
+
+}
