--- src/main/java/org/opensrp/web/utils/Constants.java
@@ -1,9 +1,0 @@
-package org.opensrp.web.utils;
-
-public class Constants {
-
-    public static final String LIMIT = "limit";
-
-    public static final int DEFAULT_LIMIT = 25;
-
-}
